// src/models/game-events.model.ts
export enum GameEvents {
  PLAYERSLIST = 'PLAYERSLIST',
  JOINPLAYER ='JOINPLAYER',
  JOINEDPLAYER ='JOINEDPLAYER',
  PLAYERREADY = 'PLAYERREADY',
  READY = 'READY',
  REGISTERPLAYER = 'REGISTERPLAYER',
  STARTGAME = 'STARTGAME',
  GAMEREGISTERPLAYERS = 'GAMEREGISTERPLAYERS',
  PLAYERROUNDREADY = 'PLAYERROUNDREADY',
  STARTROUND = 'STARTROUND',
  GAMEMOVE = 'GAMEMOVE',
  ENDROUND = 'ENDROUND',
  TOKENSINCOMING = 'TOKENSINCOMING',
  PLAYERROUNDLAUNCH = 'PLAYERROUNDLAUNCH'
}